requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
boto3==1.34.29
botocore==1.34.29
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
cssselect==1.2.0
cssselect2==0.7.0
da-vinci==0.4.0
daphne==4.1.2
datetime-truncate==1.1.1
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.6
django-admin-chart==1.0.2
django-allauth==0.61.1
django-bower==5.2.0
django-ckeditor==6.7.0
django-contrib-postgres==0.0.1
django-debug-toolbar==4.3.0
django-elasticsearch-dsl==8.0
django-jalali==6.0.1
django-js-asset==2.2.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-redis==5.4.0
django-render-partial==0.4
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-storages==1.14.2
django-thumbnails==0.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-swagger==0.2.2
drf-spectacular==0.27.2
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.13.0
filelock==3.13.1
ghasedakpack==0.1.13
hiredis==2.3.2
html5lib==1.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
itypes==1.2.0
jdatetime==4.1.1
Jinja2==3.1.3
jmespath==1.0.1
jsonpath-rw==1.4.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kavenegar==1.1.2
lxml==5.1.0
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==23.2
parsel==1.8.1
pillow==10.2.0
platformdirs==4.1.0
ply==3.11
postgres==4.0
Protego==0.3.0
psycopg==3.1.17
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.0.0
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
queuelib==1.6.2
redis==5.0.1
referencing==0.35.0
reportlab==4.0.9
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.18.0
s3transfer==0.10.0
Scrapy==2.11.1
service-identity==24.1.0
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
sorl-thumbnail==12.10.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.3.0
tldextract==5.1.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzlocal==5.2
Unidecode==1.3.8
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.7
w3lib==2.1.2
webencodings==0.5.1
xhtml2pdf==0.2.15
zeep==4.2.1
zope.interface==6.2
celery~=5.4.0